Fixes for various warnings.

+++ src/openvrml/OpenVRML/doc2.hpp
@@ -3,21 +3,21 @@
 // OpenVRML
 //
 // Copyright (C) 2000  Braden N. McDaniel
-// 
+//
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Lesser General Public
 // License as published by the Free Software Foundation; either
 // version 2.1 of the License, or (at your option) any later version.
-// 
+//
 // This library is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 // Lesser General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU Lesser General Public
 // License along with this library; if not, write to the Free Software
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
-// 
+//
 
 # ifndef OPENVRML_DOC2_HPP
 #   define OPENVRML_DOC2_HPP
@@ -59,7 +59,7 @@ namespace OpenVRML {
         Doc2(const Doc2 &);
         Doc2 & operator=(const Doc2 &);
 
-        bool filename(char * fn, int nfn);
+        bool filename(char * fn, size_t nfn);
 
 #   ifdef macintosh
         char * convertCommonToMacPath(char * fn, int nfn);
